+++ src/app/api/consoles/route.ts
@@ -1,5 +1,5 @@
 import { getVideogameConsoles } from '@/consoles/application/GetVideogameConsoles'
-import { InMemoryVideogameConsoleRepository } from '@/consoles/data/InMemoryVideogameConsoleRepository'
+import { InMemoryVideogameConsoleRepository } from '@/consoles/infra/InMemoryVideogameConsoleRepository'
 
 export async function GET(): Promise<Response> {
   const repository = InMemoryVideogameConsoleRepository.getInstance()
