+++ src/app/api/users/route.ts
@@ -1,5 +1,5 @@
 import { getUsers } from '@/users/application/GetUsers'
-import { InMemoryUserRepository } from '@/users/data/InMemoryUserRepository'
+import { InMemoryUserRepository } from '@/users/infra/InMemoryUserRepository'
 
 export async function GET(): Promise<Response> {
   const repository = InMemoryUserRepository.getInstance()
